* master: (26 commits)
  Fix processing bug with cover art script
  MBS-3643: Allow musixmatch for the lyrics whitelist
  Add IMSLP to the list of URL specializations
  More space between the two link sections
  Fix quotation marks on tag lookup results
  Properly catch NoChanges, but still die on other exceptions.
  Display release group ARs on the release relationships page
  Display release group URLs on release pages
  Split the arranger AR into generic/vocal/instrument ARs
  Correctly update artist_credit.name on artist merge
  Use Try::Tiny in Controller::AutoEditorElections.
  MBS-3719 Fix display issues with search results containing standalone recordings - alternating row colouring broken for standalone recordings - no tagger links for standalone recordings - columns not highlighted correctly for standalone recordings
  Make tagger links for recordings bring the release down into Picard; not a NAT/standalone recording.
  Replace all use of TryCatch with TryTiny.
  fix the HEAD merge thingy
  Caps changes
  introduce brackets again for release group wording
  Add explanation text + link for direct search @ Picard search page
  Add explanation text + link for direct search @ Picard search page
  Fixing it right
  ...

+++ lib/MusicBrainz/Server/Controller/AutoEditorElections.pm
@@ -2,7 +2,7 @@ package MusicBrainz::Server::Controller::AutoEditorElections;
 BEGIN { use Moose; extends 'MusicBrainz::Server::Controller' }
 
 use MusicBrainz::Server::Translation qw( l );
-use TryCatch;
+use Try::Tiny;
 
 __PACKAGE__->config( namespace => 'elections' );
 
@@ -55,9 +55,11 @@ sub _load {
         try {
             return $c->model('AutoEditorElection')->get_by_id($id);
         }
-        catch (MusicBrainz::Server::Exceptions::InvalidInput $e) {
-            $c->stash( message => $e->message );
-            $c->detach('/error_500');
+        catch {
+            if (ref ($_) eq 'MusicBrainz::Server::Exceptions::InvalidInput') {
+                $c->stash( message => $_->message );
+                $c->detach('/error_500');
+            }
         };
     }
     else {
